+++ CHANGELOG.md
@@ -3,6 +3,45 @@
 All notable changes to this project will be documented in this file. See
 [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.28.0](https://github.com/narmi/design_system/compare/v6.27.1...v6.28.0) (2026-09-02)
+
+### feat
+
+* **Callout,TimelineEvent:** add divided event; introduce callout ([d2d48e1](https://github.com/narmi/design_system/commit/d2d48e12a999876236a76962833b33f957c73f5a))
+
+## [6.27.1](https://github.com/narmi/design_system/compare/v6.27.0...v6.27.1) (2026-09-02)
+
+### fix
+
+* **Chip:** add isLabelSelectable prop to allow selecting label text ([2b2682c](https://github.com/narmi/design_system/commit/2b2682c75983f083d0f84f0f0aaceaf1441e7b27))
+
+## [6.27.0](https://github.com/narmi/design_system/compare/v6.26.0...v6.27.0) (2026-08-31)
+
+### feat
+
+* **Tabs:** allow start and end inline tab content ([8ec0f16](https://github.com/narmi/design_system/commit/8ec0f1641dc480c75819bd7b0645185c5362cdcb))
+
+### fix
+
+* **Tabs:** adjust custom content placement ([baf18b7](https://github.com/narmi/design_system/commit/baf18b757e9096fab7fea4bacdf564c59355e380))
+* extra space at startContent ([46d7574](https://github.com/narmi/design_system/commit/46d7574327b0e0d41898ba6ad20d83716a7fdfbc))
+
+## [6.26.0](https://github.com/narmi/design_system/compare/v6.25.0...v6.26.0) (2026-08-31)
+
+### ci
+
+* adjust zizmor rules and fixes ([1436428](https://github.com/narmi/design_system/commit/1436428f3710e05fcfd8e49efda20102f1433cc7))
+
+### feat
+
+* **ci:** add zizmor workflow security audit ([9796d7d](https://github.com/narmi/design_system/commit/9796d7d68363b1e6c1901ba761056e1857e3c93d))
+
+## [6.25.0](https://github.com/narmi/design_system/compare/v6.24.0...v6.25.0) (2026-08-27)
+
+### feat
+
+* **tokens:** add colorblind forced color modes ([bbaca1c](https://github.com/narmi/design_system/commit/bbaca1cb1b5cd4abbb3b215b38a2c8677739a3e1))
+
 ## [6.24.0](https://github.com/narmi/design_system/compare/v6.23.0...v6.24.0) (2026-08-21)
 
 ### feat